fix(backend): address run message pagination review
