+++ backend/tests/test_runs_api_endpoints.py
@@ -5,6 +5,7 @@ from __future__ import annotations
 from unittest.mock import AsyncMock, MagicMock
 
 from _router_auth_helpers import make_authed_test_app
+from _run_message_pagination_helpers import assert_run_message_page
 from fastapi.testclient import TestClient
 
 from app.gateway.routers import runs
@@ -109,11 +110,7 @@ def test_run_messages_default_page_keeps_newest_messages_when_extra_row_returned
         event_store=_make_event_store(rows),
     )
     with TestClient(app) as client:
-        response = client.get("/api/runs/run-2/messages")
-    assert response.status_code == 200
-    body = response.json()
-    assert body["has_more"] is True
-    assert [m["seq"] for m in body["data"]] == list(range(17, 67))
+        assert_run_message_page(client, "/api/runs/run-2/messages", expected_seq=list(range(17, 67)))
 
 
 def test_run_messages_before_seq_page_keeps_newest_side_when_extra_row_returned():
@@ -125,11 +122,11 @@ def test_run_messages_before_seq_page_keeps_newest_side_when_extra_row_returned(
         event_store=_make_event_store(rows),
     )
     with TestClient(app) as client:
-        response = client.get("/api/runs/run-2/messages?before_seq=18&limit=16")
-    assert response.status_code == 200
-    body = response.json()
-    assert body["has_more"] is True
-    assert [m["seq"] for m in body["data"]] == list(range(2, 18))
+        assert_run_message_page(
+            client,
+            "/api/runs/run-2/messages?before_seq=18&limit=16",
+            expected_seq=list(range(2, 18)),
+        )
 
 
 def test_run_messages_after_seq_page_keeps_oldest_side_when_extra_row_returned():
@@ -141,11 +138,11 @@ def test_run_messages_after_seq_page_keeps_oldest_side_when_extra_row_returned()
         event_store=_make_event_store(rows),
     )
     with TestClient(app) as client:
-        response = client.get("/api/runs/run-2/messages?after_seq=10")
-    assert response.status_code == 200
-    body = response.json()
-    assert body["has_more"] is True
-    assert [m["seq"] for m in body["data"]] == list(range(11, 61))
+        assert_run_message_page(
+            client,
+            "/api/runs/run-2/messages?after_seq=10",
+            expected_seq=list(range(11, 61)),
+        )
 
 
 def test_run_messages_passes_after_seq_to_event_store():
